Move the rest of core::unicode::char to core::unicode
